refactor: State Event Effect 분리
