+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/screen/pomodoro/rest/PomodoroRestViewModel.kt
@@ -1,15 +1,10 @@
 package com.pomonyang.mohanyang.presentation.screen.pomodoro.rest
 
-import androidx.annotation.DrawableRes
 import androidx.lifecycle.viewModelScope
 import com.mohanyang.presentation.R
-import com.pomonyang.mohanyang.data.repository.pomodoro.PomodoroTimerRepository
 import com.pomonyang.mohanyang.domain.usecase.AdjustPomodoroTimeUseCase
 import com.pomonyang.mohanyang.domain.usecase.GetSelectedPomodoroSettingUseCase
 import com.pomonyang.mohanyang.presentation.base.BaseViewModel
-import com.pomonyang.mohanyang.presentation.base.ViewEvent
-import com.pomonyang.mohanyang.presentation.base.ViewSideEffect
-import com.pomonyang.mohanyang.presentation.base.ViewState
 import com.pomonyang.mohanyang.presentation.model.setting.toModel
 import com.pomonyang.mohanyang.presentation.screen.PomodoroConstants.MAX_REST_MINUTES
 import com.pomonyang.mohanyang.presentation.screen.PomodoroConstants.MIN_REST_MINUTES
@@ -18,31 +13,10 @@ import javax.inject.Inject
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 
-data class PomodoroRestState(
-    val plusButtonSelected: Boolean = false,
-    val minusButtonSelected: Boolean = false,
-    val plusButtonEnabled: Boolean = true,
-    val minusButtonEnabled: Boolean = true
-) : ViewState
-
-sealed interface PomodoroRestEvent : ViewEvent {
-    data class OnPlusButtonClick(val isPlusButtonSelected: Boolean) : PomodoroRestEvent
-    data class OnMinusButtonClick(val isMinusButtonSelected: Boolean) : PomodoroRestEvent
-    data object OnEndPomodoroClick : PomodoroRestEvent
-    data object OnFocusClick : PomodoroRestEvent
-}
-
-sealed interface PomodoroRestEffect : ViewSideEffect {
-    data class ShowSnackbar(val message: String, @DrawableRes val iconRes: Int) : PomodoroRestEffect
-    data object GoToHome : PomodoroRestEffect
-    data object GoToPomodoroFocus : PomodoroRestEffect
-}
-
 @HiltViewModel
 class PomodoroRestViewModel @Inject constructor(
     private val getSelectedPomodoroSettingUseCase: GetSelectedPomodoroSettingUseCase,
     private val adjustPomodoroTimeUseCase: AdjustPomodoroTimeUseCase,
-    private val pomodoroTimerRepository: PomodoroTimerRepository
 ) : BaseViewModel<PomodoroRestState, PomodoroRestEvent, PomodoroRestEffect>() {
 
     init {
