+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/screen/pomodoro/waiting/PomodoroRestWaitingViewModel.kt
@@ -1,15 +1,11 @@
 package com.pomonyang.mohanyang.presentation.screen.pomodoro.waiting
 
-import androidx.annotation.DrawableRes
 import androidx.lifecycle.viewModelScope
 import com.mohanyang.presentation.BuildConfig
 import com.mohanyang.presentation.R
 import com.pomonyang.mohanyang.domain.usecase.AdjustPomodoroTimeUseCase
 import com.pomonyang.mohanyang.domain.usecase.GetSelectedPomodoroSettingUseCase
 import com.pomonyang.mohanyang.presentation.base.BaseViewModel
-import com.pomonyang.mohanyang.presentation.base.ViewEvent
-import com.pomonyang.mohanyang.presentation.base.ViewSideEffect
-import com.pomonyang.mohanyang.presentation.base.ViewState
 import com.pomonyang.mohanyang.presentation.model.setting.toModel
 import com.pomonyang.mohanyang.presentation.screen.PomodoroConstants.MAX_FOCUS_MINUTES
 import com.pomonyang.mohanyang.presentation.screen.PomodoroConstants.MIN_FOCUS_MINUTES
@@ -21,38 +17,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
-
-data class PomodoroRestWaitingState(
-    val plusButtonSelected: Boolean = false,
-    val minusButtonSelected: Boolean = false,
-    val plusButtonEnabled: Boolean = true,
-    val minusButtonEnabled: Boolean = true,
-    val forceGoHome: Boolean = false
-) : ViewState
-
-sealed interface PomodoroRestWaitingEvent : ViewEvent {
-    data class Init(
-        val exceedTime: Int,
-        val focusTime: Int,
-        val type: String
-    ) : PomodoroRestWaitingEvent
-
-    data object OnNavigationClick : PomodoroRestWaitingEvent
-    data object OnEndFocusClick : PomodoroRestWaitingEvent
-    data object OnStartRestClick : PomodoroRestWaitingEvent
-    data class OnPlusButtonClick(val isPlusButtonSelected: Boolean) : PomodoroRestWaitingEvent
-    data class OnMinusButtonClick(val isMinusButtonSelected: Boolean) : PomodoroRestWaitingEvent
-}
-
-sealed interface PomodoroRestWaitingSideEffect : ViewSideEffect {
-    data object GoToPomodoroSetting : PomodoroRestWaitingSideEffect
-    data object GoToPomodoroRest : PomodoroRestWaitingSideEffect
-
-    data class ShowSnackbar(
-        val message: String,
-        @DrawableRes val iconRes: Int
-    ) : PomodoroRestWaitingSideEffect
-}
 
 @HiltViewModel
 class PomodoroRestWaitingViewModel @Inject constructor(
